[GSKAN-281] Add postcode information on communes (when available)
There might be several communes with exactly the same name.
Until now, the select2 jquery dropdown tool was merging entries with similar
names, so we only got the first entry, the others were never proposed.
And there was no way to discriminate them.
An laternative would be to use
https://api-adresse.data.gouv.fr/search/?q=saint%20pierre%20du%20mont&lng=-0.9761&lat=49.3815
and it context information, butthen we would still need to retrieve the
id as per https://www.data.gouv.fr/fr/apidoc/#!/spatial/suggest_zones
and then again to retrieve the bbox. It would add an additionnal request level.
And t would also need to work on the select2 dropdown display to display
multiline values...